refactor: move cfg scheduler out of pag
