+++ scripts/cfg_combiner.py
@@ -42,7 +42,8 @@ class CFGCombinerScript(UIWrapper):
             cfg_dict = {
                 "denoiser": None,
                 "pag_params": None,
-                "scfg_params": None
+                "scfg_params": None,
+                "cfgi_params": None
             }
             setattr(p, 'incant_cfg_params', cfg_dict)
 
@@ -58,11 +59,13 @@ class CFGCombinerScript(UIWrapper):
             pag_active = p.extra_generation_params.get('PAG Active', False)
             cfg_active = p.extra_generation_params.get('CFG Interval Enable', False)
             scfg_active = p.extra_generation_params.get('SCFG Active', False)
+            cfgi_active = p.extra_generation_params.get('CFG Interval Enable', False)
 
             if not any([
                         pag_active,
                         cfg_active,
-                        scfg_active
+                        scfg_active,
+                        cfgi_active 
                     ]):
                 return
 
@@ -112,7 +115,8 @@ class CFGCombinerScript(UIWrapper):
                                     **kwargs,
                                     original_func = denoiser.combine_denoised_original,
                                     pag_params = cfg_dict['pag_params'],
-                                    scfg_params = cfg_dict['scfg_params']
+                                    scfg_params = cfg_dict['scfg_params'],
+                                    cfgi_params = cfg_dict['cfgi_params']
                                 )
                             patched_combine_denoised = patches.patch(__name__, denoiser, "combine_denoised", pass_conds_func)
                             setattr(denoiser, 'combine_denoised_patched', True)
@@ -158,8 +162,9 @@ def combine_denoised_pass_conds_list(*args, **kwargs):
         original_func = kwargs.get('original_func', None)
         pag_params = kwargs.get('pag_params', None)
         scfg_params = kwargs.get('scfg_params', None)
+        cfgi_params = kwargs.get('cfgi_params', None)
 
-        if pag_params is None and scfg_params is None:
+        if pag_params is None and scfg_params is None and cfgi_params is None:
                 logger.warning("No reason to hijack combine_denoised")
                 return original_func(*args)
 
@@ -173,9 +178,9 @@ def combine_denoised_pass_conds_list(*args, **kwargs):
 
                 # 1. CFG Interval
                 # Overrides cfg_scale if pag_params is not None
-                if pag_params is not None:
-                        if pag_params.cfg_interval_enable:
-                                cfg_scale = pag_params.cfg_interval_scheduled_value
+                if cfgi_params is not None:
+                        if cfgi_params.cfg_interval_enable:
+                                cfg_scale = cfgi_params.cfg_interval_scheduled_value
 
                 # 2. PAG
                 pag_x_out = None
