+++ scripts/incantation_base.py
@@ -15,6 +15,7 @@ from scripts.pag import PAGExtensionScript
 from scripts.save_attn_maps import SaveAttentionMapsScript
 from scripts.cfg_combiner import CFGCombinerScript
 from scripts.smoothed_energy_guidance import SEGExtensionScript
+from scripts.cfg_scheduler import CFGSchedulerExtensionScript
 
 logger = logging.getLogger(__name__)
 logger.setLevel(environ.get("SD_WEBUI_LOG_LEVEL", logging.INFO))
@@ -38,6 +39,7 @@ submodules: list[SubmoduleInfo] = [
         SubmoduleInfo(module=SEGExtensionScript()),
         SubmoduleInfo(module=SCFGExtensionScript()),
         SubmoduleInfo(module=PAGExtensionScript()),
+        SubmoduleInfo(module=CFGSchedulerExtensionScript()),
         SubmoduleInfo(module=T2I0ExtensionScript()),
         SubmoduleInfo(module=IncantExtensionScript()),
 ]
